Migrate handlers for intents VideosHandler and SpeakersHandler
